[REFAC] : chatService 분리 및 추상화
